backend/requirements.txt
Django==2.2.15
django-cors-headers==3.4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
psycopg2==2.8.5
